Plan, PlanHeader
A tester les pages et modaux liées aux design/plan. création, modification, etc.
